requirements.txt
requests
pandas
sqlalchemy
python-dotenv
yfinance
matplotlib
flask